requirements.txt
Django==1.10.5
djangorestframework==3.5.4
packaging==16.8
psycopg2==2.6.2
pyparsing==2.1.10
requests==2.13.0
six==1.10.0
ujson==1.35
gunicorn
